fix test dependencies
